+++ core_tools/data/ds/data_set_core.py
@@ -1,46 +1,15 @@
+from datetime import datetime
+import logging
+import time
+
 from core_tools.data.ds.data_set_DataMgr import m_param_organizer, dataset_data_description
 from core_tools.data.SQL.SQL_dataset_creator import SQL_dataset_creator
 
-import datetime
-import time
 
-class data_set_desciptor(object):
-    def __init__(self, variable, is_time=False, is_JSON=False):
-        self.var = variable
-        self.is_time = is_time
-        self.is_JSON = is_JSON
+logger = logging.getLogger(__name__)
 
-    def __get__(self, obj, objtype):
-        value = getattr(getattr(obj,"_data_set__data_set_raw"), self.var)
-        if self.is_time:
-            return datetime.datetime.fromtimestamp(value)
-
-        return value
 
 class data_set:
-    completed = data_set_desciptor('completed')
-
-    dbname = data_set_desciptor('dbname')
-    table_name = data_set_desciptor('SQL_table_name')
-    name = data_set_desciptor('exp_name')
-
-    exp_id = data_set_desciptor('exp_id')
-    exp_uuid = data_set_desciptor('exp_uuid')
-    exp_name = data_set_desciptor('exp_name')
-
-    project = data_set_desciptor('project')
-    set_up = data_set_desciptor('set_up')
-    sample_name = data_set_desciptor('sample')
-
-    metadata = data_set_desciptor('metadata')
-    snapshot = data_set_desciptor('snapshot')
-    keywords = data_set_desciptor('keywords')
-    starred = data_set_desciptor('starred')
-
-    run_timestamp = data_set_desciptor('UNIX_start_time', is_time=True)
-    run_timestamp_raw = data_set_desciptor('UNIX_start_time')
-    completed_timestamp = data_set_desciptor('UNIX_stop_time', is_time=True)
-    completed_timestamp_raw = data_set_desciptor('UNIX_stop_time')
 
     def __init__(self, ds_raw):
         self.id = None
@@ -48,6 +17,78 @@ class data_set:
         self.__repr_attr_overview = []
         self.__init_properties(m_param_organizer(ds_raw.measurement_parameters_raw))
         self.last_commit = time.time()
+
+    @property
+    def completed(self):
+        return self._data_set__data_set_raw.completed
+
+    @property
+    def dbname(self):
+        return self._data_set__data_set_raw.dbname
+
+    @property
+    def table_name(self):
+        return self._data_set__data_set_raw.SQL_table_name
+
+    @property
+    def name(self):
+        return self._data_set__data_set_raw.exp_name
+
+    @property
+    def exp_id(self):
+        return self._data_set__data_set_raw.exp_id
+
+    @property
+    def exp_uuid(self):
+        return self._data_set__data_set_raw.exp_uuid
+
+    @property
+    def exp_name(self):
+        return self._data_set__data_set_raw.exp_name
+
+    @property
+    def project(self):
+        return self._data_set__data_set_raw.project
+
+    @property
+    def set_up(self):
+        return self._data_set__data_set_raw.set_up
+
+    @property
+    def sample_name(self):
+        return self._data_set__data_set_raw.sample
+
+    @property
+    def metadata(self):
+        return self._data_set__data_set_raw.metadata
+
+    @property
+    def snapshot(self):
+        return self._data_set__data_set_raw.snapshot
+
+    @property
+    def keywords(self):
+        return self._data_set__data_set_raw.keywords
+
+    @property
+    def starred(self):
+        return self._data_set__data_set_raw.starred
+
+    @property
+    def run_timestamp(self):
+        return datetime.fromtimestamp(self._data_set__data_set_raw.UNIX_start_time)
+
+    @property
+    def run_timestamp_raw(self):
+        return self._data_set__data_set_raw.UNIX_start_time
+
+    @property
+    def completed_timestamp(self):
+        return datetime.fromtimestamp(self._data_set__data_set_raw.UNIX_stop_time)
+
+    @property
+    def completed_timestamp_raw(self):
+        return self._data_set__data_set_raw.UNIX_stop_time
 
     def __len__(self):
         return len(self.__repr_attr_overview)
@@ -79,7 +120,7 @@ class data_set:
                 if j == 0:
                     setattr(self, 'm' + str(i), ds_descript)
 
-                if j == 0 and n_sets==1: #consistent printing
+                if j == 0 and n_sets == 1:  # consistent printing
                     repr_attr_overview += [('m' + str(i), ds_descript)]
                     ds_descript.name = 'm' + str(i)
                 else:
@@ -104,7 +145,8 @@ class data_set:
         Add results to the dataset
 
         Args:
-            input_data (dict<int, list<np.ndarray>>) : dict with as key the id of the measured parameter and the data that is measured.
+            input_data (dict<int, list<np.ndarray>>):
+                dict with as key the id of the measured parameter and the data that is measured.
         '''
         for m_param in self.__data_set_raw.measurement_parameters:
             if m_param.id_info in input_data.keys():
@@ -135,12 +177,12 @@ class data_set:
         '''
         Updates dataset in case only part of the points were downloaded.
         '''
-        if self.completed == False:
+        if not self.completed:
             SQL_ds_creator = SQL_dataset_creator()
             self.completed = SQL_ds_creator.is_completed(self.exp_uuid)
             self.__data_set_raw.sync_buffers()
 
-    def __write_to_db(self, force = False):
+    def __write_to_db(self, force: bool = False):
         '''
         update values every 200ms to the database.
 
@@ -151,16 +193,22 @@ class data_set:
         # increase flush interval for long measurements to reduce overhead
         run_duration = current_time - self.__data_set_raw.UNIX_start_time
         flush_interval = 0.25
-        if run_duration > 10.0: flush_interval *= 2
-        if run_duration > 30.0: flush_interval *= 2
-        if current_time - self.last_commit > flush_interval or force==True:
+        if run_duration > 10.0:
+            flush_interval *= 2
+        if run_duration > 30.0:
+            flush_interval *= 2
+        if current_time - self.last_commit > flush_interval or force:
+            t_start = time.perf_counter()
             self.__data_set_raw.sync_buffers()
             SQL_ds_creator = SQL_dataset_creator()
             SQL_ds_creator.update_write_cursors(self.__data_set_raw)
             self.last_commit = time.time()
+            duration = time.perf_counter() - t_start
+            if duration > 0.1:
+                logger.info(f"Write to db took {duration:.3f} s")
 
     def __repr__(self):
-        output_print = "DataSet :: {}\n\nid = {}\nuuid = {}\n\n".format(self.name, self.exp_id, self.exp_uuid)
+        output_print = f"DataSet :: {self.name}\n\nid = {self.exp_id}\nuuid = {self.exp_uuid}\n\n"
         output_print += "| idn             | label           | unit     | size                     |\n"
         output_print += "---------------------------------------------------------------------------\n"
         for i in self.__repr_attr_overview:
@@ -168,9 +216,9 @@ class data_set:
                 output_print += j[1].__repr__()
                 output_print += "\n"
 
-        output_print += "set_up : {}\n".format(self.set_up)
-        output_print += "project : {}\n".format(self.project)
-        output_print += "sample_name : {}\n".format(self.sample_name)
+        output_print += f"set_up : {self.set_up}\n"
+        output_print += f"project : {self.project}\n"
+        output_print += f"sample_name : {self.sample_name}\n"
         return output_print
 
     def close(self):