+++ core_tools/data/measurement.py
@@ -243,3 +243,4 @@ class Measurement:
             return False
 
         return False
+